Pulled out OmnisciServerWorker to separate module
