+++ server_worker/__init__.py
@@ -1,0 +1,1 @@
+from .server_worker import OmnisciServerWorker
